% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiCox.R
\name{multiCox}
\alias{multiCox}
\title{Run coxPH in series}
\usage{
multiCox(rawTerms, data, event, variable)
}
\arguments{
\item{rawTerms}{- pass in output from either reduceTerms or reduceTermsPar}

\item{data}{- data set in the form of dataframe}

\item{event}{- event variable indicating censorship}

\item{variable}{- to perform regression on (e.g. Dmax)}
}
\description{
Allows the study of AIC from any number of models. Can be faster than multiCoxParallel for smaller numbers of models.
}
\examples{
multiCox()
}
\keyword{multCoxParallel}
