% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regressionPoints.R
\name{regressionPoints}
\alias{regressionPoints}
\title{Outputs text file with regression points}
\usage{
regressionPoints(coxObject, excelFile, printName)
}
\arguments{
\item{coxObject}{- pass in the coxPH (e.g. A <- coxph(), pass in "A")}

\item{excelFile}{- excelFile following format in example}

\item{printName}{- name of output text file}
}
\description{
Allows one to produced multiple modified survival curves based on input provided in an excel spreadsheet
}
\examples{
regressionPoints()
}
\keyword{regressionPoints}
